refactor: 패키지명 변경으로 인한 코드 수정
physical_profile -> physical
