+++ src/test/java/com/atwoz/member/infrastructure/member/MemberJpaRepositoryTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import static com.atwoz.member.fixture.domain.member.MemberFixture.일반_유저_생성;
+import static com.atwoz.member.fixture.MemberFixture.일반_유저_생성;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
